+++ src/main/scala/fr/janalyse/ssh/ShellOperations.scala
@@ -583,24 +583,59 @@ trait ShellOperations extends CommonOperations with LazyLogging {
   def id: String = execute("id")
   
   /**
-   * sudo without password test
-   * @return true if sudo is available and no password is required for current user
+   * Does the command "sudo su -" without password works ?
+   * The typical usage that maximizes compatibilities accross various linux is to pipe the
+   *    command to the sudo -S su -
+   * Options such as -k, -A, -p ... may not be supported everywhere. 
+   * Some notes :
+   *     BAD because we want to test the su
+   *       sudo -n echo OK 2>/dev/null
+   *
+   *     BAD because with older linux, -n option was not available
+   *       sudo -n su - -c "echo OK" 2>/dev/null
+   *
+   *     ~GOOD but NOK if only su - is allowed 
+   *       echo | sudo -S su - -c echo "OK" 2>/dev/null
+   *
+   *     GOOD
+   *       echo "echo OK" | sudo -S su - 2>/dev/null
+   * @return true if just "sudo su -" is possible without password for current user
    */
-   def sudoNoPasswordTest():Boolean = {
-    val testedmsg="SUDOOK"
-    execute(s"""echo "echo $testedmsg" | sudo -S su - 2>/dev/null""").trim.contains(testedmsg)
+  def sudoSuMinusOnlyWithoutPasswordTest():Boolean = {
+     val testedmsg="SUDOOK"
+     execute(s"""echo "echo $testedmsg" | sudo -S su - 2>/dev/null""").trim.contains(testedmsg)
   }
-   // BAD because we want to test the su
-   //   sudo -n echo OK 2>/dev/null
-   //
-   // BAD because with older linux, -n option was not available
-   //   sudo -n su - -c "echo OK" 2>/dev/null
-   //
-   // ~GOOD but NOK if only su - is allowed 
-   //   echo | sudo -S su - -c echo "OK" 2>/dev/null
-   //
-   // GOOD
-   //   echo "echo OK" | sudo -S su - 2>/dev/null
+
+  /**
+   * 
+   * Here tty can be preserved
+   */
+  def sudoSuMinusOnlyWithPasswordTest():Boolean = {
+    val cur = s"""SUDO_PROMPT="password:" sudo -S su -"""
+    // TODO
+    false
+  }
+  
+  /**
+   * Does the command sudo "su - -c theGivenCommand" works ?
+   * Transparently with or without password
+   * @return true if it works
+   */
+  def sudoSuMinusWithCommandTest(cmd:String="whoami"):Boolean = {
+    val askpass=""
+    val password=options.password.password.getOrElse("")
+    val script=
+     s"""
+        |echo '$password'
+        |""".stripMargin
+    
+    //TODO 
+        
+    execute(s"""$askpass | SUDO_PROMPT="" sudo -S su - -c $cmd 2>&1 >/dev/null; echo $$?""")
+      .trim
+      .equals("0")
+  }
+
   
   // ==========================================================================================